Score y condición de derrota

+++ Assets/Scripts/Managers/Events/GameEvents/OnEndScreenResult.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 public class OnEndScreenResult
 {
     public string Title { get; private set; }
-    public int Score { get; private set; }
+    public long Score { get; private set; }
 
-    public OnEndScreenResult(string title, int score)
+    public OnEndScreenResult(string title, long score)
     {
         Title = title;
         Score = score;
